Refactor to use composite in all components instead of useContext
