--- src/App/AppUi.js
@@ -1,79 +1,0 @@
-import React from "react";
-
-import { TodoCounter } from "../TodoCounter";
-import { TodoSearch } from "../TodoSearch";
-import { TodoList } from "../TodoList";
-import { TodoItem } from "../TodoItem";
-import { TodosLoading } from "../TodosLoading";
-import { TodosError } from "../TodosError";
-import { EmptyTodos } from "../EmptyTodos";
-import { CreateTodoButton } from "../CreateTodoButton";
-import { TodoContext } from "../TodoContext";
-import { Modal } from "../Modal";
-import { TodoForm } from "../TodoForm";
-import { TodoHeader } from "../TodoHeader";
-
-function AppUi() {
-  // Puede usarse lo de abajo de usar context que permite de forma mas prolija usar el TodoContext.consumer
-  const {
-    loading,
-    error,
-    searchedTodos,
-    completeTodo,
-    deleteTodo,
-    openModal,
-    totalTodos,
-    completedTodos,
-    searchValue,
-    setSearchValue,
-  } = React.useContext(TodoContext);
-
-  return (
-    <>
-      <TodoHeader>
-        <TodoCounter
-          loading={loading}
-          totalTodos={totalTodos}
-          completedTodos={completedTodos}
-        />
-        <TodoSearch
-          loading={loading}
-          searchValue={searchValue}
-          setSearchValue={setSearchValue}
-        />
-      </TodoHeader>
-
-      <br />
-      <TodoList>
-        {loading && (
-          <>
-            <TodosLoading />
-            <TodosLoading />
-            <TodosLoading />
-          </>
-        )}
-        {error && <TodosError />}
-        {!loading && searchedTodos.length === 0 && <EmptyTodos />}
-
-        {searchedTodos.map((todo) => (
-          <TodoItem
-            key={todo.text}
-            text={todo.text}
-            completed={todo.completed}
-            onComplete={() => completeTodo(todo.text)}
-            onDelete={() => deleteTodo(todo.text)}
-          />
-        ))}
-      </TodoList>
-
-      <CreateTodoButton />
-      {openModal && (
-        <Modal>
-          <TodoForm></TodoForm>
-        </Modal>
-      )}
-    </>
-  );
-}
-
-export { AppUi };
