+++ src/CreateTodoButton/index.js
@@ -1,9 +1,7 @@
 import React from "react";
 import "./CreateTodoButton.css";
-import { TodoContext } from "../TodoContext";
 
-function CreateTodoButton() {
-  const { setOpenModal, openModal } = React.useContext(TodoContext);
+function CreateTodoButton({ setOpenModal, openModal }) {
   return (
     <button
       className="CreateTodoButton"
